fix: downcase system names in the default template
When we create a system, it's name is an ordinary lisp symbol without
any |quoting|, so the system file name can't have camel case if we want
to recognize a system by name from the registry. Do the same with *.lisp
file, because the component is downcased in the system definition. Fixes #10.